backend/requirements.txt
# AI Quiz Generation & Attempt System — Backend Dependencies

# Core framework
django>=4.2,<5.0
djangorestframework>=3.15

# PostgreSQL adapter
psycopg2-binary>=2.9

# Environment variable management
python-dotenv>=1.0

# JWT Authentication (DFD 1.0)
djangorestframework-simplejwt>=5.3

# CORS headers (Next.js frontend support)
django-cors-headers>=4.3

# AI Integration (DFD 2.0) — choose one or both
openai>=1.0
google-generativeai>=0.5
